Move SingleLineTests into (new) navigation package
