+++ richtextfx/src/integrationTest/java/org/fxmisc/richtext/keyboard/NavigationTests.java
@@ -43,188 +43,6 @@ public class NavigationTests {
         }
     }
 
-    public class SingleLineTests extends InlineCssTextAreaAppTest {
-
-        String[] words = { "the", "cat", "can", "walk" };
-        String fullText = String.join(" ", words);
-
-        private int wordStart(int wordIndex) {
-            return entityStart(wordIndex, words);
-        }
-
-        private int wordEnd(int wordIndex) {
-            return entityEnd(wordIndex, words, area);
-        }
-
-        private void moveCaretTo(int position) {
-            area.moveTo(position);
-        }
-
-        @Override
-        public void start(Stage stage) throws Exception {
-            super.start(stage);
-            stage.setWidth(300);
-            area.replaceText(fullText);
-        }
-
-        public class NoModifiers {
-
-            @Test
-            public void left() {
-                moveCaretTo(wordStart(1));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                type(LEFT);
-
-                assertEquals(wordEnd(0), area.getCaretPosition());
-                assertTrue(area.getSelectedText().isEmpty());
-            }
-
-            @Test
-            public void right() {
-                moveCaretTo(wordStart(1));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                type(RIGHT);
-
-                assertEquals(wordStart(1) + 1, area.getCaretPosition());
-                assertTrue(area.getSelectedText().isEmpty());
-            }
-
-        }
-
-        public class ShortcutDown {
-
-            @Before
-            public void setup() {
-                press(SHORTCUT);
-            }
-
-            @Test
-            public void left() {
-                moveCaretTo(wordEnd(3));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                // first left goes to boundary of current word
-                type(LEFT);
-
-                assertEquals(wordStart(3), area.getCaretPosition());
-                assertTrue(area.getSelectedText().isEmpty());
-
-                // second left skips space and goes to end boundary of prev word
-                type(LEFT);
-
-                assertEquals(wordStart(2), area.getCaretPosition());
-                assertTrue(area.getSelectedText().isEmpty());
-            }
-
-            @Test
-            public void right() {
-                moveCaretTo(wordStart(0));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                // first right goes to boundary of current word
-                type(RIGHT);
-
-                assertEquals(wordEnd(0), area.getCaretPosition());
-                assertTrue(area.getSelectedText().isEmpty());
-
-                // second right skips space and goes to end boundary of next word
-                type(RIGHT);
-
-                assertEquals(wordEnd(1), area.getCaretPosition());
-                assertTrue(area.getSelectedText().isEmpty());
-            }
-
-            @Test
-            public void a() {
-                assertTrue(area.getSelectedText().isEmpty());
-
-                type(A);
-
-                assertEquals(area.getText(), area.getSelectedText());
-            }
-
-        }
-
-        public class ShiftDown {
-
-            @Before
-            public void setup() {
-                press(SHIFT);
-            }
-
-            @Test
-            public void left() {
-                moveCaretTo(wordStart(1));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                type(LEFT);
-
-                assertEquals(wordEnd(0), area.getCaretPosition());
-                assertEquals(" ", area.getSelectedText());
-            }
-
-            @Test
-            public void right() {
-                moveCaretTo(wordEnd(0));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                type(RIGHT);
-
-                assertEquals(wordStart(1), area.getCaretPosition());
-                assertEquals(" ", area.getSelectedText());
-            }
-
-        }
-
-        public class ShortcutShiftDown {
-
-            @Before
-            public void setup() {
-                press(SHORTCUT, SHIFT);
-            }
-
-            @Test
-            public void left() {
-                moveCaretTo(wordEnd(3));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                // first left goes to boundary of current word
-                type(LEFT);
-
-                assertEquals(wordStart(3), area.getCaretPosition());
-                assertEquals(words[3], area.getSelectedText());
-
-                // second left skips space and goes to end boundary of prev word
-                type(LEFT);
-
-                assertEquals(wordStart(2), area.getCaretPosition());
-                assertEquals(words[2] + " " + words[3], area.getSelectedText());
-            }
-
-            @Test
-            public void right() {
-                moveCaretTo(wordStart(0));
-                assertTrue(area.getSelectedText().isEmpty());
-
-                // first right goes to boundary of current word
-                type(RIGHT);
-
-                assertEquals(wordEnd(0), area.getCaretPosition());
-                assertEquals(words[0], area.getSelectedText());
-
-                // second right skips space and goes to end boundary of next word
-                type(RIGHT);
-
-                assertEquals(wordEnd(1), area.getCaretPosition());
-                assertEquals(words[0] + " " + words[1], area.getSelectedText());
-            }
-
-        }
-
-    }
-
     public class MultiLineGridlikeTextTests extends InlineCssTextAreaAppTest {
 
         public final String[] lines = {
